Moved validation out of the projects service

+++ src/projects/projects.controller.ts
@@ -14,6 +14,8 @@ import {
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger/dist';
 import { Prisma } from '@prisma/client';
+import { ActiveUser } from '../iam/decorators/active-user.decorator';
+import { ActiveUserData } from '../iam/active-user-data';
 import { QueryParams } from '../dto/query-params.dto';
 import { CreateProjectDto } from './dto/create-project.dto';
 import { ProjectsQueryParams } from './dto/projects-query-params.dto';
@@ -31,12 +33,17 @@ export class ProjectsController {
   }
 
   @Get()
-  async findAll(@Query() query: ProjectsQueryParams) {
+  async findAll(
+    @ActiveUser('sub', ParseIntPipe) userId: number,
+    @Query() query: ProjectsQueryParams,
+  ) {
     const { skip, take } = query;
     return await this.projectsService.findAll({
       skip: skip ? parseInt(skip) : QueryParams.SKIP_DEFAULT,
       take: take ? parseInt(take) : QueryParams.TAKE_DEFAULT,
-      where: {},
+      where: {
+        userId,
+      },
       orderBy: {
         name: Prisma.SortOrder.asc,
       },
@@ -44,19 +51,45 @@ export class ProjectsController {
   }
 
   @Get(':id')
-  async findOne(@Param('id', ParseIntPipe) id: number) {
-    const project = await this.projectsService.findOne(id);
+  async findOne(
+    @ActiveUser('sub', ParseIntPipe) userId: number,
+    @Param('id', ParseIntPipe) id: number,
+  ) {
+    const project = await this.projectsService.findOne({
+      where: {
+        id,
+        userId,
+      },
+      include: {
+        sections: true,
+        tasks: true,
+      },
+    });
+
     if (!project) {
       throw new NotFoundException();
     }
+
     return project;
   }
 
   @Patch(':id')
   async update(
+    @ActiveUser('sub', ParseIntPipe) userId: number,
     @Param('id', ParseIntPipe) id: number,
     @Body() updateProjectDto: UpdateProjectDto,
   ) {
+    const projectExists = await this.projectsService.exists({
+      where: {
+        id,
+        userId,
+      },
+    });
+
+    if (!projectExists) {
+      throw new NotFoundException();
+    }
+
     const project = await this.projectsService.update(id, updateProjectDto);
     if (!project) {
       throw new NotFoundException();
@@ -67,7 +100,21 @@ export class ProjectsController {
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  async remove(@Param('id', ParseIntPipe) id: number) {
+  async remove(
+    @ActiveUser('sub', ParseIntPipe) userId: number,
+    @Param('id', ParseIntPipe) id: number,
+  ) {
+    const projectExists = await this.projectsService.exists({
+      where: {
+        id,
+        userId,
+      },
+    });
+
+    if (!projectExists) {
+      throw new NotFoundException();
+    }
+
     const project = await this.projectsService.remove(id);
     if (!project) {
       throw new NotFoundException();
